Debugging snippets
Allow the snippet files to persist after compilation to help with debugging.
**Note:** Since we reuse the same snippet filename for all snippets, only the last snippet will be preserved if a document contains more than one.

+++ tex/gregoriotex.lua
@@ -995,8 +995,10 @@ local function direct_gabc(gabc, header, allow_deprecated)
     end
     glog:close()
   end
-  os.remove(snippet_filename)
-  os.remove(snippet_logname)
+  if not (debug_types_activated['snippet'] or debug_types_activated['all']) then
+    os.remove(snippet_filename)
+    os.remove(snippet_logname)
+  end
 end
 
 local function get_gregoriotexluaversion()
